게스트북 배포
조회수가계속증감됨[배포후에 쿠기가안잡힘]

+++ BG-Rest/src/main/java/com/aia/guest/service/GuestDeleteService.java
@@ -25,13 +25,14 @@ public class GuestDeleteService {
 		String uri = "/upload";
 		String realPath= req.getSession().getServletContext().getRealPath(uri);
 		
+		if(photo!=null) {
 		File file= new File(realPath,photo);
 		System.out.println("삭제경로주소 :"+file);
 		
 		 if(file.exists()) {
 			 file.delete();
 		 }
-		
+		}
 		
 		
 		return dao.deleteIdx(guest_idx);
